+++ src/main/java/com/hyunsolution/dangu/common/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.hyunsolution.dangu.common.apiResponse;
+package com.hyunsolution.dangu.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
