+++ src/main/java/com/hyunsolution/dangu/common/exception/ExceptionDto.java
@@ -1,4 +1,4 @@
-package com.hyunsolution.dangu.common.apiResponse;
+package com.hyunsolution.dangu.common.exception;
 
 import javax.validation.constraints.NotNull;
 import lombok.Getter;
